requirements.txt
llama_index==0.9.22
openai==1.6.1
python-dotenv==1.0.0
Requests==2.31.0
streamlit==1.29.0
